stats: set title in createAxis() function
Except for count-axes the title of the axis was set immediately
after creating the axis. Instead, take the title as a parameter
to createAxis(). Create a createCountAxis() wrapper that calls
createAxis() with the default title.

Since the callers do not call Qt's setTitleText() function anymore,
we can abstract away from the QtChart axis classes and use only
our wrappers in the chart code, thus paving the way to use
our own custom axes.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>